+++ tencentcloud/data_source_tc_vpn_gateway_routes.go
@@ -12,7 +12,7 @@ data "tencentcloud_vpn_gateways" "foo" {
   tags = {
     test = "tf"
   }
-}a
+}
 ```
 */
 package tencentcloud